unrefactor the PvpBtlle
